backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "mulhoo",email = "oliviamulholland94@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
